AP-6268: Error message of failing node in streaming changed
The name of the exception has been prepended to the actual node error
message when logged. As a consequence the logged error and node error
differ which causes the testflow framework to fail (since it expects
those to be the same).

AP-6268 (Components cannot be used with Try Catch nodes)